Add Depth Information script
Each of the 4 species used calcify over a range of habitats per Lakhani et al. 2022. The .mat file contains information on the distribution of ACD, and the depth realization script uses it to get a random realization from it. An example of using this script is in Figure_2 code (upcoming)